+++ app/helpers/dialog_helper.rb
@@ -1,0 +1,5 @@
+module DialogHelper
+  def button_to_close_dialog(content = nil, **, &)
+    tag.form tag.button(content, **, &), method: :dialog
+  end
+end